+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1012,7 +1012,7 @@ class NullGroupedProvidersSchema:
             (NGPcol.potential_grouped_provider, pl.Boolean()),
             (IndCQC.ascwds_filled_posts_dedup_clean, pl.Float64()),
             (IndCQC.care_home, pl.String()),
-            (IndCQC.number_of_beds, pl.Int32()),
+            (IndCQC.number_of_beds, pl.Int64()),
         ]
     )
